refactor(Captcha): cleanup code
